move common methods to request_utils
