+++ mlflow/azure/client.py
@@ -6,7 +6,7 @@ from copy import deepcopy
 import urllib
 import logging
 
-from mlflow.utils import rest_utils
+from mlflow.utils import rest_utils, request_utils
 from mlflow.utils.file_utils import read_chunk
 
 _logger = logging.getLogger(__name__)
@@ -33,7 +33,7 @@ def put_adls_file_creation(sas_url, headers):
         else:
             _logger.debug("Removed unsupported '%s' header for ADLS Gen2 Put operation", name)
 
-    with rest_utils.cloud_storage_http_request(
+    with request_utils.cloud_storage_http_request(
         "put", request_url, headers=request_headers
     ) as response:
         rest_utils.augmented_raise_for_status(response)
@@ -66,7 +66,7 @@ def patch_adls_file_upload(sas_url, local_file, start_byte, size, position, head
             _logger.debug("Removed unsupported '%s' header for ADLS Gen2 Patch operation", name)
 
     data = read_chunk(local_file, size, start_byte)
-    with rest_utils.cloud_storage_http_request(
+    with request_utils.cloud_storage_http_request(
         "patch", request_url, data=data, headers=request_headers
     ) as response:
         rest_utils.augmented_raise_for_status(response)
@@ -91,7 +91,7 @@ def patch_adls_flush(sas_url, position, headers):
         else:
             _logger.debug("Removed unsupported '%s' header for ADLS Gen2 Patch operation", name)
 
-    with rest_utils.cloud_storage_http_request(
+    with request_utils.cloud_storage_http_request(
         "patch", request_url, headers=request_headers
     ) as response:
         rest_utils.augmented_raise_for_status(response)
@@ -118,7 +118,7 @@ def put_block(sas_url, block_id, data, headers):
         else:
             _logger.debug("Removed unsupported '%s' header for Put Block operation", name)
 
-    with rest_utils.cloud_storage_http_request(
+    with request_utils.cloud_storage_http_request(
         "put", request_url, data=data, headers=request_headers
     ) as response:
         rest_utils.augmented_raise_for_status(response)
@@ -146,7 +146,7 @@ def put_block_list(sas_url, block_list, headers):
         else:
             _logger.debug("Removed unsupported '%s' header for Put Block List operation", name)
 
-    with rest_utils.cloud_storage_http_request(
+    with request_utils.cloud_storage_http_request(
         "put", request_url, data=data, headers=request_headers
     ) as response:
         rest_utils.augmented_raise_for_status(response)
